Serious refactoring

+++ src/presentation/web/src/app/components/storage/manufacturers/get/master/manufacturer-get-master.component.ts
@@ -4,7 +4,8 @@ import {Router} from '@angular/router';
 import {Subject} from 'rxjs';
 import {takeUntil} from 'rxjs/operators';
 
-import {faPen, faTimes, IconDefinition} from '@fortawesome/free-solid-svg-icons';
+import {faPen, faTimes} from '@fortawesome/free-solid-svg-icons';
+import Swal, {SweetAlertResult} from 'sweetalert2';
 
 import {ManufacturersService} from '../../../../../../api/services/manufacturers.service';
 import {ManufacturersListModel} from '../../../../../../api/models/manufacturers-list.model';
@@ -16,38 +17,67 @@ import {ManufacturersListModel} from '../../../../../../api/models/manufacturers
   providers: [ManufacturersService],
 })
 export class ManufacturerGetMasterComponent implements OnInit, OnDestroy {
-  public faPen: IconDefinition = faPen;
-  public faTimes: IconDefinition = faTimes;
-  public viewModel: ManufacturersListModel = new ManufacturersListModel();
+  public faPen = faPen;
+  public faTimes = faTimes;
+  public viewModel = new ManufacturersListModel();
   private stop$ = new Subject<void>();
 
   constructor(private titleService: Title, private manufacturersService: ManufacturersService, private router: Router) {
     this.titleService.setTitle('Manufacturers - Sample Angular');
   }
 
-  public ngOnInit(): void {
-    this.manufacturersService.getAll(50)
+  public ngOnInit() {
+    this.manufacturersService.getAll()
       .pipe(takeUntil(this.stop$))
       .subscribe(result => this.viewModel = result, error => console.error(error));
   }
 
-  public ngOnDestroy(): void {
+  public ngOnDestroy() {
     this.stop$.next();
     this.stop$.complete();
   }
 
-  public redirectToUpdate(id: number = 0): void {
+  public redirectToUpdate(id = 0) {
     this.router.navigate(['/manufacturer/update', id]);
   }
 
-  public processDelete(id: number): void {
-    this.manufacturersService.delete(id)
-      .pipe(takeUntil(this.stop$))
-      .subscribe(result => {
-        console.log(result);
-        this.viewModel.manufacturers.forEach((item, index) => {
-          if (item.manufacturerId === id) this.viewModel.manufacturers.splice(index, 1);
-        });
-      }, error => console.error(error));
+  public requestDelete(id: number) {
+    this.askToDelete().then((result: SweetAlertResult) => {
+      if (!result.value) return;
+
+      this.manufacturersService.delete(id)
+        .pipe(takeUntil(this.stop$))
+        .subscribe(() => this.processDelete(id), error => console.error(error));
+    });
+  }
+
+  private askToDelete() {
+    return Swal.fire({
+      title: 'Are you sure you want to permanently delete the record from the repository?',
+      text: 'You can’t undo this action.',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Delete entry',
+    });
+  }
+
+  private processDelete(id: number) {
+    this.viewModel.manufacturers.forEach((item, index) => {
+      if (item.manufacturerId === id) this.viewModel.manufacturers.splice(index, 1);
+    });
+
+    this.notifySuccessDelete();
+  }
+
+  private notifySuccessDelete() {
+    Swal.fire({
+      position: 'bottom-end',
+      icon: 'success',
+      title: 'Manufacturer successfully deleted',
+      showConfirmButton: false,
+      timer: 1500,
+    });
   }
 }
